pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robocloud-arm"
version = "0.1.0"
description = "RoboCloud Arm control package"
requires-python = ">=3.9"
dependencies = [
    "pyserial",
    "SpeechRecognition",
    "pyaudio",
    "opencv-python",
    "numpy",
    "readchar",
]

[tool.setuptools]
packages = ["pi", "pi.controller", "pi.input", "pi.perception"]
